@@ -6,21 +6,20 @@ import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import javax.inject.Inject;
-import javax.persistence.PersistenceException;
-import javax.ws.rs.Consumes;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-
 import cafe.model.CafeRepository;
 import cafe.model.entity.Coffee;
+import jakarta.inject.Inject;
+import jakarta.persistence.PersistenceException;
+import jakarta.ws.rs.Consumes;
+import jakarta.ws.rs.DELETE;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.POST;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.WebApplicationException;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
 
 @Path("coffees")
 public class CafeResource {
@@ -31,17 +30,17 @@ public class CafeResource {
 	private CafeRepository cafeRepository;
 
 	@GET
-	@Produces({ MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON })
+	@Produces({ MediaType.APPLICATION_JSON })
 	public List<Coffee> getAllCoffees() {
 		return this.cafeRepository.getAllCoffees();
 	}
 
 	@POST
-	@Consumes({ MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON })
+	@Consumes({ MediaType.APPLICATION_JSON })
 	public Response createCoffee(Coffee coffee) {
 		try {
 			coffee = this.cafeRepository.persistCoffee(coffee);
-			return Response.created(URI.create("/" + coffee.getId())).build();
+			return Response.created(URI.create(coffee.getId().toString())).build();
 		} catch (PersistenceException e) {
 			logger.log(Level.SEVERE, "Error creating coffee {0}: {1}.", new Object[] { coffee, e });
 			throw new WebApplicationException(e, Response.Status.INTERNAL_SERVER_ERROR);
@@ -50,7 +49,7 @@ public class CafeResource {
 
 	@GET
 	@Path("{id}")
-	@Produces({ MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON })
+	@Produces({ MediaType.APPLICATION_JSON })
 	public Coffee getCoffeeById(@PathParam("id") Long coffeeId) {
 		return this.cafeRepository.findCoffeeById(coffeeId);
 	}
@@ -61,8 +60,7 @@ public class CafeResource {
 		try {
 			this.cafeRepository.removeCoffeeById(coffeeId);
 		} catch (IllegalArgumentException ex) {
-			logger.log(Level.SEVERE, "Error calling deleteCoffee() for coffeeId {0}: {1}.",
-					new Object[] { coffeeId, ex });
+			logger.log(Level.SEVERE, "Error deleting coffee with ID {0}: {1}.", new Object[] { coffeeId, ex });
 			throw new WebApplicationException(Response.Status.NOT_FOUND);
 		}
 	}
